% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\docType{data}
\name{genotype}
\alias{genotype}
\title{Genotype in Numeric Format}
\format{
A data frame with 410 rows and 4906 variables.
}
\usage{
genotype
}
\description{
Genotypic data of 410 maize hybrids with 4906 SNPs.
}
\keyword{datasets}
